backend_intranet/requirements.txt
Django==5.2.4
Pillow>=10.0.0
python-decouple>=3.8
whitenoise>=6.6.0
gunicorn>=21.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5